chore: refactor ajax
