+++ .changeset/rude-frogs-run.md
@@ -1,0 +1,10 @@
+---
+'@lion/ajax': minor
+---
+
+Update public API:
+
+- AjaxClient is now Ajax
+- AjaxClientFetchError is now AjaxFetchError
+- request and requestJson methods of Ajax class are renamed as fetch and fetchJson respectively
+- getCookie and validateOptions is not part of the public API any more
